requirements.txt
distributed==1.25.0
docutils==0.14
Flask==1.0.2
Flask-Cors==3.0.7
gunicorn==19.0.0
h5py==2.8.0
imagesize==1.1.0
importlib-metadata==0.6
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
Jinja2==2.10
jsonschema==2.6.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keyring==16.1.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
nbconvert==5.4.0
nbformat==4.4.0
numpy==1.15.4
numpydoc==0.8.0
packaging==18.0
pandas==0.23.4
path.py==11.5.0
pickleshare==0.7.5
Pillow==5.3.0
requests==2.20.1
ruamel-yaml==0.15.46
scikit-image==0.14.0
scikit-learn==0.20.1
scipy==1.1.0
seaborn==0.9.0
tensorboard==1.12.0
tensorflow==1.12.0
urllib3==1.23
Werkzeug==0.14.1
widgetsnbextension==3.4.2
